feat: add meatballs menu (#3781)
